+++ spring-boot-kafka-avro/src/main/java/com/sivalabs/springbootkafkaavro/repository/PersonRepository.java
@@ -1,8 +1,9 @@
 package com.sivalabs.springbootkafkaavro.repository;
 
 import com.sivalabs.springbootkafkaavro.entity.PersonEntity;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PersonRepository extends JpaRepository<PersonEntity, Integer> {}
+public interface PersonRepository extends JpaRepository<PersonEntity, Long> {}
